debug

+++ 3FT/ft.c
@@ -602,7 +602,7 @@ int FT_stat(char *path, boolean *type, size_t *length) {
         temp = Node_getFileContents(queryNode);
         fileContents = (char*)DynArray_get(temp, 0);
         type = (boolean*)TRUE;
-        length = (int*)strlen(fileContents);
+        length = (size_t*)strlen(fileContents);
     }
     else {
         type = (boolean*)FALSE;
